Split up slow pq test
